+++ docs/Measurements.Client/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,13 @@
 # Measurements.Client Release Notes
 
+## Version 1.1.4
+
+- Bugfix. Don't reference Measurements from Client.
+
+## Version 1.1.3
+
+- Bugfix. Deserializing MeasurementDto
+
 ## Version 1.1.2
 
 - Serialize response with string enum values.